aphrodite-web/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==5.0.1
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
six==1.17.0
waitress==3.0.2
Werkzeug==3.1.3
